Convert to double explicitly

+++ elastic4s-client-pekko/src/main/scala/com/sksamuel/elastic4s/pekko/DefaultBlacklist.scala
@@ -34,7 +34,7 @@ private[pekko] class DefaultBlacklist(min: FiniteDuration, max: FiniteDuration, 
       val retries = record.retries + 1
 
       val untilTime = now + Math
-        .min(min.toNanos * Math.pow(2, retries * 0.5), max.toNanos)
+        .min(min.toNanos * Math.pow(2, retries * 0.5), max.toNanos.toDouble)
         .toLong
 
       val updatedRecord =
